adds explicit deps, fixes imageData refs, moves templating
