+++ index.html
@@ -1,15 +1,7 @@
 <!DOCTYPE html>
 <html>
 <head>
-  <title>video tripz</title>
-  <style>
-    video {
-      display:none;
-    }
-    canvas {
-      background: transparent;
-    }
-  </style>
+  <title>vidz</title>
   <link rel="stylesheet" href="css/styles.css">
 </head>
 <body>
@@ -75,7 +67,6 @@
 
     </div>
   </div>
-  <script src='js/dot.js'></script>
   <script src="build.js"></script>
 </body>
 </html>
