+++ js/models/brightness-filter.js
@@ -31,7 +31,7 @@ const BrightnessFilter = VL.Model.extend({
       data[ii+2] = this.lookup[data[ii+2]];
     }
 
-    return data;
+    return imageData;
   },
 
   _buildLookup: function(value) {
