+++ js/models/inversion-filter.js
@@ -24,7 +24,7 @@ const InversionFilter = VL.Model.extend({
       data[ii + 2] = this.lookup[data[ii + 2]];
     }
 
-    return data;
+    return imageData;
   },
 
   _buildLookup: function() {
