+++ js/models/threshold-filter.js
@@ -26,7 +26,7 @@ const ThresholdFilter = VL.Model.extend({
       g = data[ii+1];
       b = data[ii+2];
 
-      var gray = ((3*r+4*g+b) >>> 3);
+      var gray = ((3*r+4*g+b) >> 3);
 
       if (threshold < 0) {
         data[ii] = data[ii+1] = data[ii+2] = gray >= (~threshold+1) ? 0 : 255;
@@ -35,7 +35,7 @@ const ThresholdFilter = VL.Model.extend({
       }
     }
 
-    return data;
+    return imageData;
   }
 });
