+++ js/models/xor-filter.js
@@ -28,7 +28,7 @@ const XORFilter = VL.Model.extend({
       data[ii+2] = b ^ (data[ii * 2] || 255);
     }
 
-    return data
+    return imageData;
 	}
 });
