--- js/withWorker.js
@@ -1,5 +1,0 @@
-var withWorker = function(object, pathToWorkerFile) {
-  object.worker = new Worker(pathToWorkerFile);
-
-  return object;
-};
